Safer errgroup (#8)
This PR introduces a new package that's forked from Go's `errgroup`,
but:
1. **Requires the usage of `context`**: it's impossible to use a `Group`
without passing a context. The `Group` then keeps the context
internally, and pass it to the functions provided to the `Go` and
`TryGo` methods;
2. **Recovers from `panic`s**: this design decision was made because we
envision that this package will be used in HTTP/gRPC handlers, or
similar setups, where programs already usually recover from panics that
happen in the scope of a request, but panics happening in spawned
goroutines would _still_ take down the whole process if unhandled.

TODOs:
* [ ] add a `README`
* [ ] add a `Makefile`
* [ ] add a package-level godoc
* [ ] use a different name?

---------

Co-authored-by: Scott Blum <dragonsinth@gmail.com>